--- teal/binary_search.lua
@@ -1,24 +1,0 @@
-local _tl_compat; if (tonumber((_VERSION or ''):match('[%d.]*$')) or 0) < 5.3 then local p, m = pcall(require, 'compat53.module'); if p then _tl_compat = m end end; local math = _tl_compat and _tl_compat.math or math; local function binary_search(list, item, cmp)
-   local len = #list
-   local mid
-   local s, e = 1, len
-   while s <= e do
-      mid = math.floor((s + e) / 2)
-      local val = list[mid]
-      local res = cmp(val, item)
-      if res then
-         if mid == len then
-            return mid, val
-         else
-            if not cmp(list[mid + 1], item) then
-               return mid, val
-            end
-         end
-         s = mid + 1
-      else
-         e = mid - 1
-      end
-   end
-end
-
-return binary_search
