+++ teal/type_reporter.lua
@@ -1,0 +1,539 @@
+local _tl_compat; if (tonumber((_VERSION or ''):match('[%d.]*$')) or 0) < 5.3 then local p, m = pcall(require, 'compat53.module'); if p then _tl_compat = m end end; local assert = _tl_compat and _tl_compat.assert or assert; local ipairs = _tl_compat and _tl_compat.ipairs or ipairs; local pairs = _tl_compat and _tl_compat.pairs or pairs; local string = _tl_compat and _tl_compat.string or string; local table = _tl_compat and _tl_compat.table or table; local types = require("teal.types")
+
+
+
+
+
+
+
+
+
+
+
+
+local show_type = types.show_type
+
+local parser = require("teal.parser")
+
+
+local variables = require("teal.variables")
+
+
+local util = require("teal.util")
+local binary_search = util.binary_search
+local sorted_keys = util.sorted_keys
+
+local type_reporter = { TypeCollector = { Symbol = {} }, TypeInfo = {}, TypeReport = {}, TypeReporter = {} }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+local TypeReporter = type_reporter.TypeReporter
+
+
+
+
+
+
+
+
+
+
+
+
+local typecodes = {
+
+   NIL = 0x00000001,
+   NUMBER = 0x00000002,
+   BOOLEAN = 0x00000004,
+   STRING = 0x00000008,
+   TABLE = 0x00000010,
+   FUNCTION = 0x00000020,
+   USERDATA = 0x00000040,
+   THREAD = 0x00000080,
+
+   INTEGER = 0x00010002,
+   ENUM = 0x00010004,
+   EMPTY_TABLE = 0x00000008,
+   ARRAY = 0x00010008,
+   RECORD = 0x00020008,
+   MAP = 0x00040008,
+   TUPLE = 0x00080008,
+   INTERFACE = 0x00100008,
+   SELF = 0x00200008,
+   POLY = 0x20000020,
+   UNION = 0x40000000,
+
+   NOMINAL = 0x10000000,
+   TYPE_VARIABLE = 0x08000000,
+
+   ANY = 0xffffffff,
+   UNKNOWN = 0x80008000,
+   INVALID = 0x80000000,
+}
+
+
+
+
+
+local typename_to_typecode = {
+   ["typevar"] = typecodes.TYPE_VARIABLE,
+   ["typearg"] = typecodes.TYPE_VARIABLE,
+   ["unresolved_typearg"] = typecodes.TYPE_VARIABLE,
+   ["unresolvable_typearg"] = typecodes.TYPE_VARIABLE,
+   ["function"] = typecodes.FUNCTION,
+   ["array"] = typecodes.ARRAY,
+   ["map"] = typecodes.MAP,
+   ["tupletable"] = typecodes.TUPLE,
+   ["interface"] = typecodes.INTERFACE,
+   ["self"] = typecodes.SELF,
+   ["record"] = typecodes.RECORD,
+   ["enum"] = typecodes.ENUM,
+   ["boolean"] = typecodes.BOOLEAN,
+   ["string"] = typecodes.STRING,
+   ["nil"] = typecodes.NIL,
+   ["thread"] = typecodes.THREAD,
+   ["userdata"] = typecodes.USERDATA,
+   ["number"] = typecodes.NUMBER,
+   ["integer"] = typecodes.INTEGER,
+   ["union"] = typecodes.UNION,
+   ["nominal"] = typecodes.NOMINAL,
+   ["circular_require"] = typecodes.NOMINAL,
+   ["boolean_context"] = typecodes.BOOLEAN,
+   ["emptytable"] = typecodes.EMPTY_TABLE,
+   ["unresolved_emptytable_value"] = typecodes.EMPTY_TABLE,
+   ["poly"] = typecodes.POLY,
+   ["any"] = typecodes.ANY,
+   ["unknown"] = typecodes.UNKNOWN,
+   ["invalid"] = typecodes.INVALID,
+
+   ["none"] = typecodes.UNKNOWN,
+   ["tuple"] = typecodes.UNKNOWN,
+   ["literal_table_item"] = typecodes.UNKNOWN,
+   ["typedecl"] = typecodes.UNKNOWN,
+   ["generic"] = typecodes.UNKNOWN,
+   ["*"] = typecodes.UNKNOWN,
+}
+
+local skip_types = {
+   ["none"] = true,
+   ["tuple"] = true,
+   ["literal_table_item"] = true,
+}
+
+
+local function mark_array(x)
+   local arr = x
+   arr[0] = false
+   return x
+end
+
+function type_reporter.new()
+   local self = setmetatable({
+      next_num = 1,
+      typeid_to_num = {},
+      typename_to_num = {},
+      tr = {
+         by_pos = {},
+         types = {},
+         symbols_by_file = {},
+         globals = {},
+      },
+   }, { __index = TypeReporter })
+
+   local names = {}
+   for name, _ in pairs(types.simple_types) do
+      table.insert(names, name)
+   end
+   table.sort(names)
+
+   for _, name in ipairs(names) do
+      local ti = {
+         t = assert(typename_to_typecode[name]),
+         str = name,
+      }
+      local n = self.next_num
+      self.typename_to_num[name] = n
+      self.tr.types[n] = ti
+      self.next_num = self.next_num + 1
+   end
+
+   return self
+end
+
+function TypeReporter:store_function(ti, rt)
+   local args = {}
+   for _, fnarg in ipairs(rt.args.tuple) do
+      table.insert(args, mark_array({ self:get_typenum(fnarg), nil }))
+   end
+   ti.args = mark_array(args)
+   local rets = {}
+   for _, fnarg in ipairs(rt.rets.tuple) do
+      table.insert(rets, mark_array({ self:get_typenum(fnarg), nil }))
+   end
+   ti.rets = mark_array(rets)
+   ti.vararg = not not rt.args.is_va
+   ti.varret = not not rt.rets.is_va
+end
+
+function TypeReporter:get_typenum(t)
+
+   local n = self.typename_to_num[t.typename]
+   if n then
+      return n
+   end
+
+   assert(t.typeid)
+
+   n = self.typeid_to_num[t.typeid]
+   if n then
+      return n
+   end
+
+   local tr = self.tr
+
+
+   n = self.next_num
+
+   local rt = t
+   if rt.typename == "tuple" and #rt.tuple == 1 then
+      rt = rt.tuple[1]
+   end
+
+   if rt.typename == "typedecl" then
+      return self:get_typenum(rt.def)
+   end
+
+
+   if rt.typename == "generic" then
+      rt = rt.t
+   end
+
+   local ti = {
+      t = assert(typename_to_typecode[rt.typename]),
+      str = show_type(t, true),
+      file = t.f,
+      y = t.y,
+      x = t.x,
+   }
+   tr.types[n] = ti
+   self.typeid_to_num[t.typeid] = n
+   self.next_num = self.next_num + 1
+
+   if t.typename == "nominal" then
+      if t.found then
+         ti.ref = self:get_typenum(t.found)
+      end
+      if t.resolved then
+         rt = t
+      end
+   end
+   assert(not (rt.typename == "typedecl"))
+
+   if rt.fields then
+
+      local r = {}
+      for _, k in ipairs(rt.field_order) do
+         local v = rt.fields[k]
+         r[k] = self:get_typenum(v)
+      end
+      ti.fields = r
+      if rt.meta_fields then
+
+         local m = {}
+         for _, k in ipairs(rt.meta_field_order) do
+            local v = rt.meta_fields[k]
+            m[k] = self:get_typenum(v)
+         end
+         ti.meta_fields = m
+      end
+   end
+
+   if rt.elements then
+      ti.elements = self:get_typenum(rt.elements)
+   end
+
+   if rt.typename == "map" then
+      ti.keys = self:get_typenum(rt.keys)
+      ti.values = self:get_typenum(rt.values)
+   elseif rt.typename == "enum" then
+      ti.enums = mark_array(sorted_keys(rt.enumset))
+   elseif rt.typename == "function" then
+      self:store_function(ti, rt)
+   elseif rt.types then
+      local tis = {}
+      for _, pt in ipairs(rt.types) do
+         table.insert(tis, self:get_typenum(pt))
+      end
+      ti.types = mark_array(tis)
+   end
+
+   return n
+end
+
+function TypeReporter:add_field(rtype, fname, ftype)
+   local n = self:get_typenum(rtype)
+   local ti = self.tr.types[n]
+   assert(ti.fields)
+   ti.fields[fname] = self:get_typenum(ftype)
+end
+
+function TypeReporter:set_ref(nom, resolved)
+   local n = self:get_typenum(nom)
+   local ti = self.tr.types[n]
+   ti.ref = self:get_typenum(resolved)
+end
+
+function TypeReporter:get_collector(filename)
+   local collector = {
+      filename = filename,
+      symbol_list = {},
+   }
+
+   local ft = {}
+   self.tr.by_pos[filename] = ft
+
+   local symbol_list = collector.symbol_list
+   local symbol_list_n = 0
+
+   collector.store_type = function(y, x, typ)
+      if not typ or skip_types[typ.typename] then
+         return
+      end
+
+      local yt = ft[y]
+      if not yt then
+         yt = {}
+         ft[y] = yt
+      end
+
+      yt[x] = self:get_typenum(typ)
+   end
+
+   collector.reserve_symbol_list_slot = function(node)
+      symbol_list_n = symbol_list_n + 1
+      node.symbol_list_slot = symbol_list_n
+   end
+
+   collector.add_to_symbol_list = function(node, name, t)
+      if not node then
+         return
+      end
+      local slot
+      if node.symbol_list_slot then
+         slot = node.symbol_list_slot
+      else
+         symbol_list_n = symbol_list_n + 1
+         slot = symbol_list_n
+      end
+      symbol_list[slot] = { y = node.y, x = node.x, name = name, typ = t }
+   end
+
+   collector.begin_symbol_list_scope = function(node)
+      symbol_list_n = symbol_list_n + 1
+      symbol_list[symbol_list_n] = { y = node.y, x = node.x, name = "@{" }
+   end
+
+   collector.rollback_symbol_list_scope = function()
+      while symbol_list[symbol_list_n].name ~= "@{" do
+         symbol_list[symbol_list_n] = nil
+         symbol_list_n = symbol_list_n - 1
+      end
+   end
+
+   collector.end_symbol_list_scope = function(node)
+      if symbol_list[symbol_list_n].name == "@{" then
+         symbol_list[symbol_list_n] = nil
+         symbol_list_n = symbol_list_n - 1
+      else
+         symbol_list_n = symbol_list_n + 1
+         symbol_list[symbol_list_n] = { y = assert(node.yend), x = assert(node.xend), name = "@}" }
+      end
+   end
+
+   return collector
+end
+
+function TypeReporter:store_result(collector, globals)
+   local tr = self.tr
+
+   local filename = collector.filename
+   local symbol_list = collector.symbol_list
+
+   tr.by_pos[filename][0] = nil
+
+
+   do
+      local n = 0
+      local p = 0
+      local n_stack, p_stack = {}, {}
+      local level = 0
+      for i, s in ipairs(symbol_list) do
+         if s.typ then
+            n = n + 1
+         elseif s.name == "@{" then
+            level = level + 1
+            n_stack[level], p_stack[level] = n, p
+            n, p = 0, i
+         else
+            if n == 0 then
+               symbol_list[p].skip = true
+               s.skip = true
+            end
+            n, p = n_stack[level], p_stack[level]
+            level = level - 1
+         end
+      end
+   end
+
+   local symbols = mark_array({})
+   tr.symbols_by_file[filename] = symbols
+
+
+   do
+      local stack = {}
+      local level = 0
+      local i = 0
+      for _, s in ipairs(symbol_list) do
+         if not s.skip then
+            i = i + 1
+            local id
+            if s.typ then
+               id = self:get_typenum(s.typ)
+            elseif s.name == "@{" then
+               level = level + 1
+               stack[level] = i
+               id = -1
+            else
+               local other = stack[level]
+               level = level - 1
+               symbols[other][4] = i
+               id = other - 1
+            end
+            local sym = mark_array({ s.y, s.x, s.name, id })
+            table.insert(symbols, sym)
+         end
+      end
+   end
+
+   local gkeys = sorted_keys(globals)
+   for _, name in ipairs(gkeys) do
+      if name:sub(1, 1) ~= "@" then
+         local var = globals[name]
+         tr.globals[name] = self:get_typenum(var.t)
+      end
+   end
+
+   if not tr.symbols then
+      tr.symbols = tr.symbols_by_file[filename]
+   end
+end
+
+function TypeReporter:get_report()
+   return self.tr
+end
+
+
+
+
+
+
+function type_reporter.symbols_in_scope(tr, y, x, filename)
+   local function find(symbols, at_y, at_x)
+      local function le(a, b)
+         return a[1] < b[1] or
+         (a[1] == b[1] and a[2] <= b[2])
+      end
+      return binary_search(symbols, { at_y, at_x }, le) or 0
+   end
+
+   local ret = {}
+
+   local symbols = tr.symbols_by_file[filename]
+   if not symbols then
+      return ret
+   end
+
+   local n = find(symbols, y, x)
+
+   while n >= 1 do
+      local s = symbols[n]
+      local symbol_name = s[3]
+      if symbol_name == "@{" then
+         n = n - 1
+      elseif symbol_name == "@}" then
+         n = s[4]
+      else
+         if ret[symbol_name] == nil then
+            ret[symbol_name] = s[4]
+         end
+         n = n - 1
+      end
+   end
+
+   return ret
+end
+
+return type_reporter
