+++ teal/util.tl
@@ -1,0 +1,39 @@
+--------------------------------------------------------------------------------
+-- General algorithm utilities module, with no other dependencies
+--------------------------------------------------------------------------------
+local record util
+end
+
+function util.binary_search<T, U>(list: {T}, item: U, cmp: function(T, U): boolean): integer, T
+   local len <const> = #list
+   local mid: integer
+   local s, e = 1, len
+   while s <= e do
+      mid = math.floor((s + e) / 2)
+      local val <const> = list[mid]
+      local res <const> = cmp(val, item)
+      if res then
+         if mid == len then
+            return mid, val
+         else
+            if not cmp(list[mid + 1], item) then
+               return mid, val
+            end
+         end
+         s = mid + 1
+      else
+         e = mid - 1
+      end
+   end
+end
+
+function util.sorted_keys<A,B>(m: {A:B}):{A}
+   local keys = {}
+   for k, _ in pairs(m) do
+      table.insert(keys, k)
+   end
+   table.sort(keys)
+   return keys
+end
+
+return util
