+++ teal/variables.lua
@@ -1,0 +1,44 @@
+local types = require("teal.types")
+
+
+
+local parser = require("teal.parser")
+
+
+
+local variables = { Variable = {}, Scope = {} }
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+return variables
